requirements.txt

# Requirements for the project

streamlit
streamlit-lottie
python-dotenv
PyMuPDF
requests
